Refactor: move editor to RaiseObjectionDialog
