+++ src/modules/dashboard/components/ActionsPage/StakingWidget/StakingWidget.tsx
@@ -3,13 +3,6 @@ import { defineMessages } from 'react-intl';
 import { bigNumberify } from 'ethers/utils';
 import * as yup from 'yup';
 import { Decimal } from 'decimal.js';
-import { useEditor } from '@tiptap/react';
-import CharacterCount from '@tiptap/extension-character-count';
-import Color from '@tiptap/extension-color';
-import Placeholder from '@tiptap/extension-placeholder';
-import TextStyle from '@tiptap/extension-text-style';
-import Underline from '@tiptap/extension-underline';
-import StarterKit from '@tiptap/starter-kit';
 
 import { ActionForm } from '~core/Fields';
 import Button from '~core/Button';
@@ -66,8 +59,6 @@ const validationSchema = yup.object({
   amount: yup.number(),
 });
 
-const LIMIT = 4000;
-
 const StakingWidget = ({
   colony,
   colony: { colonyAddress, nativeTokenAddress },
@@ -106,20 +97,6 @@ const StakingWidget = ({
 
   const openRaiseObjectionDialog = useDialog(RaiseObjectionDialog);
 
-  const editor = useEditor({
-    extensions: [
-      StarterKit,
-      Underline,
-      TextStyle,
-      Color,
-      CharacterCount.configure({ limit: LIMIT }),
-      Placeholder.configure({
-        emptyEditorClass: 'is-editor-empty',
-        placeholder: 'What would you like to say?',
-      }),
-    ],
-  });
-
   const handleRaiseObjection = useCallback(
     (userHasPermission: boolean, stakingAmounts: StakingAmounts) =>
       openRaiseObjectionDialog({
@@ -128,10 +105,8 @@ const StakingWidget = ({
         canUserStake: userHasPermission,
         scrollToRef,
         ...stakingAmounts,
-        editor,
-        limit: LIMIT,
       }),
-    [colony, openRaiseObjectionDialog, scrollToRef, motionId, editor],
+    [colony, openRaiseObjectionDialog, scrollToRef, motionId],
   );
 
   const getDecimalStake = useCallback(
